refactor(docker-tests): split runner and extract SLP helpers
Phase 8.3 architectural improvements:

- Extract SLP code from helpers/utxo.rs into helpers/slp.rs
  - BchDockerOps, SLP_TOKEN_ID, SLP_TOKEN_OWNERS
  - get_prefilled_slp_privkey(), get_slp_token_id()

- Split runner.rs into per-chain setup modules
  - runner/mod.rs: Core orchestrator
  - runner/utxo.rs: MYCOIN/MYCOIN1 setup
  - runner/slp.rs: FORSLP setup
  - runner/qtum.rs: Qtum/QRC20 setup
  - runner/geth.rs: ETH/ERC20 setup
  - runner/zcoin.rs: ZCoin/Zombie setup
  - runner/tendermint.rs: Cosmos/IBC setup
  - runner/sia.rs: Sia setup

- Push cfg gates to module boundaries
  - Per-chain modules conditionally compiled via #[cfg(...)]
  - Setup calls in setup_or_reuse_nodes() match module gates
