+++ mm2src/mm2_main/tests/docker_tests/helpers/utxo.rs
@@ -56,36 +56,6 @@ use coins::utxo::UtxoActivationParams;
 ))]
 use mm2_core::mm_ctx::{MmArc, MmCtxBuilder};
 
-// UtxoCommonOps - needed for my_public_key() in SLP initialization
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use coins::utxo::UtxoCommonOps;
-
-// Transaction trait - needed for tx_hex() in SLP initialization
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use coins::Transaction;
-
-// SLP-specific imports
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use chain::TransactionOutput;
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use coins::utxo::bch::{bch_coin_with_priv_key, BchActivationRequest, BchCoin};
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use coins::utxo::slp::{slp_genesis_output, SlpOutput, SlpToken};
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use coins::utxo::utxo_common::send_outputs_from_my_address;
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use common::block_on_f01;
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use keys::{AddressBuilder, KeyPair, NetworkPrefix as CashAddrPrefix};
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use primitives::hash::H256;
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use script::Builder;
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use std::convert::TryFrom;
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-use std::sync::Mutex;
-
 // rmd160_from_priv imports
 #[cfg(any(feature = "docker-tests-ordermatch", feature = "docker-tests-swaps-utxo"))]
 use bitcrypto::dhash160;
@@ -128,19 +98,6 @@ fn get_funding_lock(ticker: &str) -> &'static AsyncMutex<()> {
         "QTUM" | "QICK" | "QORTY" => &QTUM_LOCK,
         _ => panic!("No funding lock defined for ticker: {}", ticker),
     }
-}
-
-// =============================================================================
-// SLP token metadata (SLP-only)
-// =============================================================================
-
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-lazy_static! {
-    /// SLP token ID (genesis tx hash)
-    pub static ref SLP_TOKEN_ID: Mutex<H256> = Mutex::new(H256::default());
-    /// Private keys supplied with 1000 SLP tokens on tests initialization.
-    /// Due to the SLP protocol limitations only 19 outputs (18 + change) can be sent in one transaction.
-    pub static ref SLP_TOKEN_OWNERS: Mutex<Vec<[u8; 32]>> = Mutex::new(Vec::with_capacity(18));
 }
 
 // =============================================================================
@@ -220,131 +177,6 @@ impl UtxoAssetDockerOps {
 }
 
 // =============================================================================
-// BchDockerOps (SLP features only)
-// =============================================================================
-
-/// Docker operations for BCH/SLP coins (FORSLP).
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-pub struct BchDockerOps {
-    #[allow(dead_code)]
-    ctx: MmArc,
-    coin: BchCoin,
-}
-
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-impl BchDockerOps {
-    /// Create BchDockerOps from ticker.
-    pub fn from_ticker(ticker: &str) -> BchDockerOps {
-        let conf =
-            json!({"coin": ticker,"asset": ticker,"txfee":1000,"network": "regtest","txversion":4,"overwintered":1});
-        let req = json!({"method":"enable", "bchd_urls": [], "allow_slp_unsafe_conf": true});
-        let priv_key = Secp256k1Secret::from("809465b17d0a4ddb3e4c69e8f23c2cabad868f51f8bed5c765ad1d6516c3306f");
-        let ctx = MmCtxBuilder::new().into_mm_arc();
-        let params = BchActivationRequest::from_legacy_req(&req).unwrap();
-
-        let coin = block_on(bch_coin_with_priv_key(
-            &ctx,
-            ticker,
-            &conf,
-            params,
-            CashAddrPrefix::SlpTest,
-            priv_key,
-        ))
-        .unwrap();
-        BchDockerOps { ctx, coin }
-    }
-
-    /// Initialize SLP tokens.
-    pub fn initialize_slp(&self) {
-        fill_address(&self.coin, &self.coin.my_address().unwrap(), 100000.into(), 30);
-        let mut slp_privkeys = vec![];
-
-        let slp_genesis_op_ret = slp_genesis_output("ADEXSLP", "ADEXSLP", None, None, 8, None, 1000000_00000000);
-        let slp_genesis = TransactionOutput {
-            value: self.coin.as_ref().dust_amount,
-            script_pubkey: Builder::build_p2pkh(&self.coin.my_public_key().unwrap().address_hash().into()).to_bytes(),
-        };
-
-        let mut bch_outputs = vec![slp_genesis_op_ret, slp_genesis];
-        let mut slp_outputs = vec![];
-
-        for _ in 0..18 {
-            let key_pair = KeyPair::random_compressed();
-            let address = AddressBuilder::new(
-                Default::default(),
-                Default::default(),
-                self.coin.as_ref().conf.address_prefixes.clone(),
-                None,
-            )
-            .as_pkh_from_pk(*key_pair.public())
-            .build()
-            .expect("valid address props");
-
-            block_on_f01(
-                self.native_client()
-                    .import_address(&address.to_string(), &address.to_string(), false),
-            )
-            .unwrap();
-
-            let script_pubkey = Builder::build_p2pkh(&key_pair.public().address_hash().into());
-
-            bch_outputs.push(TransactionOutput {
-                value: 1000_00000000,
-                script_pubkey: script_pubkey.to_bytes(),
-            });
-
-            slp_outputs.push(SlpOutput {
-                amount: 1000_00000000,
-                script_pubkey: script_pubkey.to_bytes(),
-            });
-            slp_privkeys.push(*key_pair.private_ref());
-        }
-
-        let slp_genesis_tx = block_on_f01(send_outputs_from_my_address(self.coin.clone(), bch_outputs)).unwrap();
-        let confirm_payment_input = ConfirmPaymentInput {
-            payment_tx: slp_genesis_tx.tx_hex(),
-            confirmations: 1,
-            requires_nota: false,
-            wait_until: wait_until_sec(30),
-            check_every: 1,
-        };
-        block_on_f01(self.coin.wait_for_confirmations(confirm_payment_input)).unwrap();
-
-        let adex_slp = SlpToken::new(
-            8,
-            "ADEXSLP".into(),
-            <&[u8; 32]>::try_from(slp_genesis_tx.tx_hash_as_bytes().as_slice())
-                .unwrap()
-                .into(),
-            self.coin.clone(),
-            1,
-        )
-        .unwrap();
-
-        let tx = block_on(adex_slp.send_slp_outputs(slp_outputs)).unwrap();
-        let confirm_payment_input = ConfirmPaymentInput {
-            payment_tx: tx.tx_hex(),
-            confirmations: 1,
-            requires_nota: false,
-            wait_until: wait_until_sec(30),
-            check_every: 1,
-        };
-        block_on_f01(self.coin.wait_for_confirmations(confirm_payment_input)).unwrap();
-        *SLP_TOKEN_OWNERS.lock().unwrap() = slp_privkeys;
-        *SLP_TOKEN_ID.lock().unwrap() = <[u8; 32]>::try_from(slp_genesis_tx.tx_hash_as_bytes().as_slice())
-            .unwrap()
-            .into();
-    }
-}
-
-#[cfg(any(feature = "docker-tests-slp", feature = "docker-tests-integration"))]
-impl CoinDockerOps for BchDockerOps {
-    fn rpc_client(&self) -> &UtxoRpcClientEnum {
-        &self.coin.as_ref().rpc_client
-    }
-}
-
-// =============================================================================
 // Docker node helpers
 // =============================================================================
 
@@ -403,18 +235,6 @@ pub fn rmd160_from_priv(privkey: Secp256k1Secret) -> H160 {
     let secret = SecretKey::from_slice(privkey.as_slice()).unwrap();
     let public = PublicKey::from_secret_key(&Secp256k1::new(), &secret);
     dhash160(&public.serialize())
-}
-
-/// Get a prefilled SLP privkey from the pool.
-#[cfg(feature = "docker-tests-slp")]
-pub fn get_prefilled_slp_privkey() -> [u8; 32] {
-    SLP_TOKEN_OWNERS.lock().unwrap().remove(0)
-}
-
-/// Get the SLP token ID as hex string.
-#[cfg(feature = "docker-tests-slp")]
-pub fn get_slp_token_id() -> String {
-    hex::encode(SLP_TOKEN_ID.lock().unwrap().as_slice())
 }
 
 /// Import an address to the coin's wallet.
