+++ mm2src/mm2_main/tests/docker_tests/runner/slp.rs
@@ -1,0 +1,26 @@
+//! SLP/BCH (FORSLP) setup for docker tests.
+
+use super::{DockerTestMode, DockerTestRunner};
+
+use crate::docker_tests::helpers::docker_ops::setup_utxo_conf_for_compose;
+use crate::docker_tests::helpers::env::KDF_FORSLP_SERVICE;
+use crate::docker_tests::helpers::slp::BchDockerOps;
+use crate::docker_tests::helpers::utxo::utxo_asset_docker_node;
+
+pub(super) fn setup(runner: &mut DockerTestRunner) {
+    match runner.config.mode {
+        DockerTestMode::Testcontainers => {
+            let node = utxo_asset_docker_node("FORSLP", 10000);
+            let for_slp_ops = BchDockerOps::from_ticker("FORSLP");
+            crate::docker_tests::helpers::docker_ops::CoinDockerOps::wait_ready(&for_slp_ops, 4);
+            for_slp_ops.initialize_slp();
+            runner.hold(node);
+        },
+        DockerTestMode::ComposeInit => {
+            setup_utxo_conf_for_compose("FORSLP", KDF_FORSLP_SERVICE);
+            let for_slp_ops = BchDockerOps::from_ticker("FORSLP");
+            crate::docker_tests::helpers::docker_ops::CoinDockerOps::wait_ready(&for_slp_ops, 4);
+            for_slp_ops.initialize_slp();
+        },
+    }
+}
